[New] Added test cases for pages admin user groups list
